requirements.txt
wheel
flask==1.1.2
gunicorn==20.0.4
openstacksdk==0.46.0
